feat: add cog enrich driver
