+++ conf/download_drivers.yaml
@@ -66,10 +66,12 @@ drivers:
     class_name: extensions.aproc.proc.download.drivers.impl.zarr_download
     configuration:
       input:
-        type: $APROC_ZARR_INPUT_STORAGE_TYPE
-        bucket: $APROC_ZARR_INPUT_STORAGE_BUCKET
+        type: $APROC_INPUT_STORAGE_TYPE
+        bucket: $APROC_INPUT_STORAGE_BUCKET
         api_key:
-          project_id: $APROC_ZARR_INPUT_STORAGE_API_KEY_PROJECT
-          private_key_id: $APROC_ZARR_INPUT_STORAGE_API_KEY_PRIVATE_KEY_ID
-          private_key: $APROC_ZARR_INPUT_STORAGE_API_KEY_PRIVATE_KEY
+          project_id: $APROC_INPUT_STORAGE_API_KEY_PROJECT
+          private_key_id: $APROC_INPUT_STORAGE_API_KEY_PRIVATE_KEY_ID
+          private_key: $APROC_INPUT_STORAGE_API_KEY_PRIVATE_KEY
+        headers:
+        #  TODO: be able to configure them
     priority: 1
