+++ docker-compose.yaml
@@ -175,11 +175,11 @@ services:
       - DOWNLOAD_S3_SECRET_ACCESS_KEY=${DOWNLOAD_S3_SECRET_ACCESS_KEY:-airssecret}
       - DOWNLOAD_S3_ASSET_HTTP_ENDPOINT_URL=${DOWNLOAD_S3_ASSET_HTTP_ENDPOINT_URL:-http://minio:9000/{}/{}}
       - CLEAN_DOWNLOAD_OUTBOX_DIR=${CLEAN_DOWNLOAD_OUTBOX_DIR:-True}
-      - APROC_ZARR_INPUT_STORAGE_TYPE=${APROC_ZARR_INPUT_STORAGE_TYPE}
-      - APROC_ZARR_INPUT_STORAGE_BUCKET=${APROC_ZARR_INPUT_STORAGE_BUCKET}
-      - APROC_ZARR_INPUT_STORAGE_API_KEY_PROJECT=${APROC_ZARR_INPUT_STORAGE_API_KEY_PROJECT}
-      - APROC_ZARR_INPUT_STORAGE_API_KEY_PRIVATE_KEY_ID=${APROC_ZARR_INPUT_STORAGE_API_KEY_PRIVATE_KEY_ID}
-      - APROC_ZARR_INPUT_STORAGE_API_KEY_PRIVATE_KEY=${APROC_ZARR_INPUT_STORAGE_API_KEY_PRIVATE_KEY}
+      - APROC_INPUT_STORAGE_TYPE=${APROC_INPUT_STORAGE_TYPE}
+      - APROC_INPUT_STORAGE_BUCKET=${APROC_INPUT_STORAGE_BUCKET}
+      - APROC_INPUT_STORAGE_API_KEY_PROJECT=${APROC_INPUT_STORAGE_API_KEY_PROJECT}
+      - APROC_INPUT_STORAGE_API_KEY_PRIVATE_KEY_ID=${APROC_INPUT_STORAGE_API_KEY_PRIVATE_KEY_ID}
+      - APROC_INPUT_STORAGE_API_KEY_PRIVATE_KEY=${APROC_INPUT_STORAGE_API_KEY_PRIVATE_KEY}
     volumes:
       - ./test/inputs:/inputs:ro
       - ./outbox/:/outbox
@@ -242,11 +242,11 @@ services:
       - APROC_INDEX_LOGIN=${APROC_INDEX_LOGIN:-None}
       - APROC_INDEX_PWD=${APROC_INDEX_PWD:-None}
       - APROC_RESOURCE_ID_HASH_STARTS_AT=${APROC_RESOURCE_ID_HASH_STARTS_AT}
-      - APROC_ZARR_INPUT_STORAGE_TYPE=${APROC_ZARR_INPUT_STORAGE_TYPE}
-      - APROC_ZARR_INPUT_STORAGE_BUCKET=${APROC_ZARR_INPUT_STORAGE_BUCKET}
-      - APROC_ZARR_INPUT_STORAGE_API_KEY_PROJECT=${APROC_ZARR_INPUT_STORAGE_API_KEY_PROJECT}
-      - APROC_ZARR_INPUT_STORAGE_API_KEY_PRIVATE_KEY_ID=${APROC_ZARR_INPUT_STORAGE_API_KEY_PRIVATE_KEY_ID}
-      - APROC_ZARR_INPUT_STORAGE_API_KEY_PRIVATE_KEY=${APROC_ZARR_INPUT_STORAGE_API_KEY_PRIVATE_KEY}
+      - APROC_INPUT_STORAGE_TYPE=${APROC_INPUT_STORAGE_TYPE}
+      - APROC_INPUT_STORAGE_BUCKET=${APROC_INPUT_STORAGE_BUCKET}
+      - APROC_INPUT_STORAGE_API_KEY_PROJECT=${APROC_INPUT_STORAGE_API_KEY_PROJECT}
+      - APROC_INPUT_STORAGE_API_KEY_PRIVATE_KEY_ID=${APROC_INPUT_STORAGE_API_KEY_PRIVATE_KEY_ID}
+      - APROC_INPUT_STORAGE_API_KEY_PRIVATE_KEY=${APROC_INPUT_STORAGE_API_KEY_PRIVATE_KEY}
     volumes:
       - ./test/inputs:/inputs:ro
     ports:
