+++ extensions/aproc/proc/download/drivers/impl/met_file.py
@@ -1,6 +1,5 @@
 import os
 from airs.core.models.model import Item, Role, ItemFormat, AssetFormat
-from aproc.core.settings import Configuration
 from extensions.aproc.proc.download.drivers.driver import Driver as DownloadDriver
 
 from extensions.aproc.proc.download.drivers.impl.utils import make_raw_archive_zip
@@ -12,7 +11,7 @@ class Driver(DownloadDriver):
 
     # Implements drivers method
     @staticmethod
-    def init(configuration: Configuration):
+    def init(configuration: dict):
         ...
 
     # Implements drivers method
@@ -23,7 +22,6 @@ class Driver(DownloadDriver):
                item_format == ItemFormat.terrasar.value or \
                item_format == ItemFormat.spot5.value
 
-    
     # Implements drivers method
     def fetch_and_transform(self, item: Item, target_directory: str, crop_wkt: str, target_projection: str, target_format: str, raw_archive: bool):
         asset = item.assets.get(Role.metadata.value)
