+++ extensions/aproc/proc/download/drivers/impl/zarr_download.py
@@ -1,70 +1,22 @@
-import enum
 import json
 import os
 import re
 import tarfile
 import tempfile
 import zipfile
-from typing import Annotated, Literal, Union
 from urllib.parse import urlparse
 
-from pydantic import BaseModel, Field, computed_field
+from pydantic import Field
 
 from airs.core.models.model import AssetFormat, Item, Role
 from extensions.aproc.proc.download.drivers.driver import \
     Driver as DownloadDriver
 from extensions.aproc.proc.download.drivers.exceptions import DriverException
 from extensions.aproc.proc.download.drivers.impl.utils import extract
+from extensions.aproc.proc.s3_configuration import S3Configuration
 
 
-class GoogleStorageConstants(str, enum.Enum):
-    AUTH_URI = "https://accounts.google.com/o/oauth2/auth"
-    TOKEN_URI = "https://oauth2.googleapis.com/token"
-    AUTH_PROVIDER_CERT_URL = "https://www.googleapis.com/oauth2/v1/certs"
-
-
-class GoogleStorageApiKey(BaseModel):
-    type: Literal["service_account"] = "service_account"
-    project_id: str
-    private_key_id: str
-    private_key: str
-    client_id: str | None = Field(None)
-    auth_uri: Literal[GoogleStorageConstants.AUTH_URI] = GoogleStorageConstants.AUTH_URI.value
-    token_uri: Literal[GoogleStorageConstants.TOKEN_URI] = GoogleStorageConstants.TOKEN_URI.value
-    auth_provider_x509_cert_url: Literal[GoogleStorageConstants.AUTH_PROVIDER_CERT_URL] = GoogleStorageConstants.AUTH_PROVIDER_CERT_URL.value
-    universe_domain: Literal["googleapis.com"] = "googleapis.com"
-
-    @computed_field
-    @property
-    def client_x509_cert_url(self) -> str:
-        return f"https://www.googleapis.com/robot/v1/metadata/x509/{self.project_id}%40appspot.gserviceaccount.com"
-
-    @computed_field
-    @property
-    def client_email(self) -> str:
-        return f"{self.project_id}@appspot.gserviceaccount.com"
-
-
-class GoogleStorage(BaseModel):
-    type: Literal["gs"] = "gs"
-    bucket: str
-    api_key: GoogleStorageApiKey
-
-
-class HttpsStorage(BaseModel):
-    type: Literal["https"] = "https"
-    headers: dict[str, str]
-
-
-class NoStorage(BaseModel):
-    type: Literal[None] = None
-
-
-Storage = Annotated[Union[GoogleStorage, HttpsStorage, NoStorage], Field(discriminator="type")]
-
-
-class Configuration(BaseModel):
-    input: Storage | None = Field(None)
+class Configuration(S3Configuration):
     chunk_size: int = Field(1000)
 
 
@@ -77,8 +29,8 @@ class Driver(DownloadDriver):
     # Implements drivers method
     def supports(item: Item) -> bool:
         data = item.assets.get(Role.data.value)
-        return item.properties.constellation == "Sentinel-2" \
-            and item.properties.processing__level == "L1C" \
+        return item.properties.constellation.lower() == "Sentinel-2".lower() \
+            and item.properties.processing__level.lower() == "L1C".lower() \
             and data is not None and data.href is not None and (
                 # data.href.startswith("s3://")
                 data.href.startswith("gs://")
@@ -105,7 +57,7 @@ class Driver(DownloadDriver):
 
         asset_href = item.assets.get(Role.data.value).href
 
-        with smart_open.open(asset_href, "rb", transport_params=self.__get_storage_parameters(asset_href)) as fb:
+        with smart_open.open(asset_href, "rb", transport_params=self.configuration.get_storage_parameters(asset_href, Driver.LOGGER)) as fb:
             with zipfile.ZipFile(fb) as raster_zip:
                 file_names = raster_zip.namelist()
                 raster_files = list(filter(
@@ -212,36 +164,5 @@ class Driver(DownloadDriver):
                 credentials = f.name
 
             return rasterio.session.GSSession(credentials)
-
-        raise NotImplementedError(f"Storage '{storage_type}' not compatible")
-
-    def __get_storage_parameters(self, href: str) -> dict:
-        storage_type = urlparse(href).scheme
-
-        if not storage_type or storage_type == "file" or storage_type == "http":
-            return {}
-
-        if storage_type == "https":
-            if self.configuration.input.type != "https":
-                Driver.LOGGER.warning("No headers is configured for HTTPS requests. Using no headers")
-                return {}
-            return {"headers": self.configuration.input.headers}
-
-        if storage_type == "gs":
-            from google.cloud.storage import Client
-            from google.oauth2 import service_account
-
-            if self.configuration.input.type != "gs":
-                import google.auth.credentials
-
-                Driver.LOGGER.warning("No api_key is configured for Google Storage, but requesting an item on Google Storage. Using anonymous credentials")
-                credentials = google.auth.credentials.AnonymousCredentials()
-            else:
-                # TODO: check if bucket match ?
-                api_key = self.configuration.input.api_key
-                credentials = service_account.Credentials.from_service_account_info(api_key)
-
-            client = Client("APROC", credentials=credentials)
-            return {"client": client}
 
         raise NotImplementedError(f"Storage '{storage_type}' not compatible")
