+++ test/env.sh
@@ -73,11 +73,11 @@ export APROC_INDEX_ENDPOINT_URL=http://elasticsearch:9200
 export APROC_INDEX_NAME=aproc_downloads
 export APROC_RESOURCE_ID_HASH_STARTS_AT=1
 
-export APROC_ZARR_INPUT_STORAGE_TYPE=""
-export APROC_ZARR_INPUT_STORAGE_BUCKET=""
-export APROC_ZARR_INPUT_STORAGE_API_KEY_PROJECT=""
-export APROC_ZARR_INPUT_STORAGE_API_KEY_PRIVATE_KEY_ID=""
-export APROC_ZARR_INPUT_STORAGE_API_KEY_PRIVATE_KEY=""
+export APROC_INPUT_STORAGE_TYPE=""
+export APROC_INPUT_STORAGE_BUCKET=""
+export APROC_INPUT_STORAGE_API_KEY_PROJECT=""
+export APROC_INPUT_STORAGE_API_KEY_PRIVATE_KEY_ID=""
+export APROC_INPUT_STORAGE_API_KEY_PRIVATE_KEY=""
 
 # AGATE
 export AGATE_PREFIX=/arlas/agate
